requirements.txt
pytorch
pandas
numpy
tdqm
einops
